refs #1: Refactor policy/{interfaces|common}.go
